requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.19.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
openpyxl>=3.1.2
numpy>=1.26.0

